Update README.md
This simple project allows you to add, view, and update job applications. It stores the job applications in a list and saves them to a JSON file when you choose to quit the program. You can expand upon this project by adding more features and functionality as per your job requirements.